+++ src/components/review-list/index.tsx
@@ -1,4 +1,4 @@
-import Review from '../review';
+import Review from '../review-component';
 import { Review as ReviewType } from '../../store/reviews/types';
 import React from 'react';
